Make some variables which are never changed final
In making the variables final we prevent the possibility of them ever being changed (outside of abnormal situations such as reflection), hence making it easier for an enterprise developer to follow along with the code, as they know that the final variables' values will never be changed and therefore do not have to look for changes to the variables in order to find out why odd behaviour is occurring. Obviously, this is largely hypothetical and just to put developers' minds at rest as there is no way that odd behaviour could ever occur in such excellently written enterprise code.