Revert "Merge pull request #208 from ThomasRooney/master"
This reverts commit 152fc6038507e6cf77705afdf6ce40f7b9300c89, reversing
changes made to 2facdda298aeba9b06153efc0c6ee95ec8deddaf.

We are very fortunate that our continuous integration software did not pick up on this merge as the QA department apparently missed how this did not build AT ALL !!
This is absolutely unacceptable, and we will be taking swift action, as soon as we get over the very busy holiday period.